Code is kind of a mess, check description
This commit was dedicated to fixing errors in the tests. I'm currently
stuck on figuring out the null cases for all of the methods, so if I do
figure that out that will be my next commit. Please ask me if you don't
understand anything that I wrote and I will do my best to explain haha